Changes for Google Cloud Function/function-source/requirements.txt
pyodbc
httplib2
oauthlib
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
google-cloud-storage
datetime
mysql.connector
pytz
beautifulsoup4